feat: adding a radius to filter the places to be displayed

+++ arlocationbased/src/main/java/com/dewakoding/arlocationbased/model/Place.kt
@@ -18,11 +18,11 @@ class Place(
     y: Float? = null,
     altitude: Double? = null,
     description: String? = null,
-    distance: Double? = null,
+    distance: Float? = null,
     photoUrl: String? = null
 ): ARPoint(id, lat, lon, x, y) {
     var name: String = name
     var description: String? = description
-    var distance: Double? = distance
+    var distance: Float? = distance
     var photoUrl: String? = photoUrl
 }
